+++ PutJedaja/src/com/java/usenko/comparator/RoomByPriceComparator.java
@@ -7,6 +7,6 @@ import java.util.Comparator;
 public class RoomByPriceComparator implements Comparator<Room> {
     @Override
     public int compare(Room o1, Room o2) {
-        return o1.getPrice() - o2.getPrice();
+        return ((int) (o1.getPrice() - o2.getPrice()));
     }
 }
